refactor: move entry points to root
