+++ src/initApp.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { App } from './app';
+import { App } from './components/app/app';
 
 ReactDOM.render(<App />, document.getElementById('appRoot'));
