+++ src/initError.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { Error } from './Error';
+import { Error } from './components/error/Error';
 
 ReactDOM.render(<Error />, document.getElementById('errorRoot'));
