+++ src/initLogin.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { Login } from './Login';
+import { Login } from './components/login/Login';
 
 ReactDOM.render(<Login />, document.getElementById('loginRoot'));
